requirement.txt
discord==1.5.1
python-dotenv==0.14.0
